Add missing project files.
