+++ Tests/MapboxNavigationTests/Snapshot++.swift
